+++ src/rr-iter.h
@@ -88,13 +88,13 @@ _getdns_rr_iter_section(_getdns_rr_iter *i)
 {
 	return !i->pkt ? (i->nxt - i->rr_type == 4 ? SECTION_QUESTION
 	                                           : SECTION_ANSWER  )
-             : i->n < GLDNS_QDCOUNT(i->pkt) ? SECTION_QUESTION
-	     : i->n < GLDNS_QDCOUNT(i->pkt)
+             : i->n < (size_t)GLDNS_QDCOUNT(i->pkt) ? SECTION_QUESTION
+	     : i->n < (size_t)GLDNS_QDCOUNT(i->pkt)
 	            + GLDNS_ANCOUNT(i->pkt) ? SECTION_ANSWER
-	     : i->n < GLDNS_QDCOUNT(i->pkt)
+	     : i->n < (size_t)GLDNS_QDCOUNT(i->pkt)
 	            + GLDNS_ANCOUNT(i->pkt)
 	            + GLDNS_NSCOUNT(i->pkt) ? SECTION_AUTHORITY
-	     : i->n < GLDNS_QDCOUNT(i->pkt)
+	     : i->n < (size_t)GLDNS_QDCOUNT(i->pkt)
 	            + GLDNS_ANCOUNT(i->pkt)
 	            + GLDNS_NSCOUNT(i->pkt)
 	            + GLDNS_ARCOUNT(i->pkt) ? SECTION_ADDITIONAL
